--- a/pkg/parquetquery/iters.go
+++ b/pkg/parquetquery/iters.go
@@ -18,6 +18,10 @@ import (
 	"go.opentelemetry.io/otel/trace"
 )
 
+type NestedValuesReader interface {
+	ReadNestedValues([]pq.Value) (int, error)
+}
+
 // RowNumber is the sequence of row numbers uniquely identifying a value
 // in a tree of nested columns, starting at the top-level and including
 // another row number for each level of nesting. -1 is a placeholder
@@ -788,20 +792,23 @@ type SyncIterator struct {
 	filter     Predicate
 
 	// Status
-	span            trace.Span
-	curr            RowNumber
-	currRowGroup    pq.RowGroup
-	currRowGroupMin RowNumber
-	currRowGroupMax RowNumber
-	currChunk       *ColumnChunkHelper
-	currPage        pq.Page
-	currPageMin     RowNumber
-	currPageMax     RowNumber
-	currValues      pq.ValueReader
-	currBuf         []pq.Value
-	currBufN        int
-	currPageN       int
-	at              IteratorResult // Current value pointed at by iterator. Returned by call Next and SeekTo, valid until next call.
+	span             trace.Span
+	curr             RowNumber
+	currRowGroup     pq.RowGroup
+	currRowGroupMin  RowNumber
+	currRowGroupMax  RowNumber
+	currChunk        *ColumnChunkHelper
+	currPage         pq.Page
+	currPageMin      RowNumber
+	currPageMax      RowNumber
+	currValues       pq.ValueReader
+	currNestedValues NestedValuesReader
+	currBuf          []pq.Value
+	currBufN         int
+	currPageN        int
+	at               IteratorResult // Current value pointed at by iterator. Returned by call Next and SeekTo, valid until next call.
+
+	defLvl byte
 
 	intern   bool
 	interner *intern.Interner
@@ -809,7 +816,7 @@ type SyncIterator struct {
 
 var _ Iterator = (*SyncIterator)(nil)
 
-func NewSyncIterator(ctx context.Context, rgs []pq.RowGroup, column int, columnName string, readSize int, filter Predicate, selectAs string, opts ...SyncIteratorOpt) *SyncIterator {
+func NewSyncIterator(ctx context.Context, rgs []pq.RowGroup, column int, columnName string, readSize int, filter Predicate, selectAs string, defLvl byte, opts ...SyncIteratorOpt) *SyncIterator {
 	// Assign row group bounds.
 	// Lower bound is inclusive
 	// Upper bound is exclusive, points at the first row of the next group
@@ -851,6 +858,7 @@ func NewSyncIterator(ctx context.Context, rgs []pq.RowGroup, column int, columnN
 		filter:     filter,
 		curr:       EmptyRowNumber(),
 		at:         at,
+		defLvl:     defLvl,
 	}
 
 	// Apply options
@@ -1088,7 +1096,7 @@ func (c *SyncIterator) seekWithinPage(to RowNumber, definitionLevel int) {
 	pq.Release(c.currPage)
 	c.currPage = pg
 	c.currPageMin = c.curr
-	c.currValues = pg.Values()
+	c.setValues(pg.Values())
 	c.currPageN = 0
 	syncIteratorPoolPut(c.currBuf)
 	c.currBuf = nil
@@ -1141,7 +1149,16 @@ func (c *SyncIterator) next() (RowNumber, *pq.Value, error) {
 		}
 		if c.currBufN >= len(c.currBuf) || len(c.currBuf) == 0 {
 			c.currBuf = c.currBuf[:cap(c.currBuf)]
-			n, err := c.currValues.ReadValues(c.currBuf)
+
+			var n int
+			var err error
+
+			if c.currNestedValues != nil {
+				n, err = c.currNestedValues.ReadNestedValues(c.currBuf)
+			} else {
+				n, err = c.currValues.ReadValues(c.currBuf)
+			}
+
 			if err != nil && !errors.Is(err, io.EOF) {
 				return EmptyRowNumber(), nil, err
 			}
@@ -1156,11 +1173,32 @@ func (c *SyncIterator) next() (RowNumber, *pq.Value, error) {
 
 		// Consume current buffer until empty
 		for c.currBufN < len(c.currBuf) {
-			v := &c.currBuf[c.currBufN]
+			var r, d byte
+			var v *pq.Value
+
+			// jpe - what about that one interaction where filter == nil, do we just do the old way?
+
+			if c.currNestedValues != nil { // jpe - better check
+
+				// if currNestedValues is set advance c.curr AS LONG AS the the definition level DOES NOT equal c.definitionLevel. in this case the value is a NULL and we can skip it
+				r := c.currPage.RepetitionLevels()[c.currPageN]
+				d := c.currPage.DefinitionLevels()[c.currPageN]
+				if d != c.defLvl {
+					// this is a null!
+					c.curr.Next(int(r), int(d))
+					c.currPageN++
+					continue
+				}
+				v = &c.currBuf[c.currBufN]
+			} else {
+				v = &c.currBuf[c.currBufN]
+				r = byte(v.RepetitionLevel())
+				d = byte(v.DefinitionLevel())
+			}
 
 			// Inspect all values to track the current row number,
 			// even if the value is filtered out next.
-			c.curr.Next(v.RepetitionLevel(), v.DefinitionLevel())
+			c.curr.Next(int(r), int(d))
 			c.currBufN++
 			c.currPageN++
 
@@ -1192,6 +1230,7 @@ func (c *SyncIterator) setPage(pg pq.Page) {
 
 	// Reset value buffers
 	c.currValues = nil
+	c.currNestedValues = nil
 	c.currPageMax = EmptyRowNumber()
 	c.currPageMin = EmptyRowNumber()
 	c.currBufN = 0
@@ -1211,7 +1250,18 @@ func (c *SyncIterator) setPage(pg pq.Page) {
 		c.currPage = pg
 		c.currPageMin = c.curr
 		c.currPageMax = rn
-		c.currValues = pg.Values()
+
+		c.setValues(pg.Values())
+	}
+}
+
+func (c *SyncIterator) setValues(v pq.ValueReader) {
+	c.currValues = v
+
+	if _, ok := v.(NestedValuesReader); ok {
+		c.currNestedValues = v.(NestedValuesReader)
+	} else {
+		c.currNestedValues = nil
 	}
 }
 
@@ -1293,7 +1343,7 @@ func NewColumnIterator(ctx context.Context, rgs []pq.RowGroup, column int, colum
 		rgs:      rgs,
 		col:      column,
 		colName:  columnName,
-		filter:   &InstrumentedPredicate{pred: filter},
+		filter:   &InstrumentedPredicate{Pred: filter},
 		selectAs: selectAs,
 		quit:     make(chan struct{}),
 		ch:       make(chan *columnIteratorBuffer, 1),