--- a/pkg/parquetquery/iters.go
+++ b/pkg/parquetquery/iters.go
@@ -18,10 +18,6 @@ import (
 	"go.opentelemetry.io/otel/trace"
 )
 
-type NestedValuesReader interface {
-	ReadNestedValues([]pq.Value) (int, error)
-}
-
 // RowNumber is the sequence of row numbers uniquely identifying a value
 // in a tree of nested columns, starting at the top-level and including
 // another row number for each level of nesting. -1 is a placeholder
@@ -132,7 +128,7 @@ func (t *RowNumber) Valid() bool {
 // null   | 1 | 1 | {  0,  1, -1, -1 }
 // gb     | 1 | 3 | {  0,  2,  0,  0 }
 // null   | 0 | 1 | {  1,  0, -1, -1 }
-func (t *RowNumber) Next(repetitionLevel, definitionLevel int) {
+/*func (t *RowNumber) Next(repetitionLevel, definitionLevel int) {
 	t[repetitionLevel]++
 
 	// the following is nextSlow() unrolled
@@ -550,11 +546,11 @@ func (t *RowNumber) Next(repetitionLevel, definitionLevel int) {
 			panicWhenInvalidDefinitionLevel(definitionLevel)
 		}
 	}
-}
+}*/
 
 // nextSlow is the original implementation of next. it is kept to test against
 // the unrolled version above
-func (t *RowNumber) nextSlow(repetitionLevel, definitionLevel int) {
+func (t *RowNumber) Next(repetitionLevel, definitionLevel, maxDefinitionLevel int) {
 	t[repetitionLevel]++
 
 	// New children up through the definition level
@@ -563,7 +559,7 @@ func (t *RowNumber) nextSlow(repetitionLevel, definitionLevel int) {
 	}
 
 	// // Children past the definition level are undefined
-	for i := definitionLevel + 1; i < len(t); i++ {
+	for i := definitionLevel + 1; i < len(t) && i <= maxDefinitionLevel; i++ {
 		t[i] = -1
 	}
 }
@@ -792,23 +788,22 @@ type SyncIterator struct {
 	filter     Predicate
 
 	// Status
-	span             trace.Span
-	curr             RowNumber
-	currRowGroup     pq.RowGroup
-	currRowGroupMin  RowNumber
-	currRowGroupMax  RowNumber
-	currChunk        *ColumnChunkHelper
-	currPage         pq.Page
-	currPageMin      RowNumber
-	currPageMax      RowNumber
-	currValues       pq.ValueReader
-	currNestedValues NestedValuesReader
-	currBuf          []pq.Value
-	currBufN         int
-	currPageN        int
-	at               IteratorResult // Current value pointed at by iterator. Returned by call Next and SeekTo, valid until next call.
+	span            trace.Span
+	curr            RowNumber
+	currRowGroup    pq.RowGroup
+	currRowGroupMin RowNumber
+	currRowGroupMax RowNumber
+	currChunk       *ColumnChunkHelper
+	currPage        pq.Page
+	currPageMin     RowNumber
+	currPageMax     RowNumber
+	currValues      pq.ValueReader
+	currBuf         []pq.Value
+	currBufN        int
+	currPageN       int
+	at              IteratorResult // Current value pointed at by iterator. Returned by call Next and SeekTo, valid until next call.
 
-	defLvl byte
+	maxDefinitionLevel int
 
 	intern   bool
 	interner *intern.Interner
@@ -816,7 +811,7 @@ type SyncIterator struct {
 
 var _ Iterator = (*SyncIterator)(nil)
 
-func NewSyncIterator(ctx context.Context, rgs []pq.RowGroup, column int, columnName string, readSize int, filter Predicate, selectAs string, defLvl byte, opts ...SyncIteratorOpt) *SyncIterator {
+func NewSyncIterator(ctx context.Context, rgs []pq.RowGroup, column int, columnName string, readSize int, filter Predicate, selectAs string, maxDefinitionLevel int, opts ...SyncIteratorOpt) *SyncIterator {
 	// Assign row group bounds.
 	// Lower bound is inclusive
 	// Upper bound is exclusive, points at the first row of the next group
@@ -848,17 +843,17 @@ func NewSyncIterator(ctx context.Context, rgs []pq.RowGroup, column int, columnN
 
 	// Create the iterator
 	i := &SyncIterator{
-		span:       span,
-		column:     column,
-		columnName: columnName,
-		rgs:        rgs,
-		readSize:   readSize,
-		rgsMin:     rgsMin,
-		rgsMax:     rgsMax,
-		filter:     filter,
-		curr:       EmptyRowNumber(),
-		at:         at,
-		defLvl:     defLvl,
+		span:               span,
+		column:             column,
+		columnName:         columnName,
+		rgs:                rgs,
+		readSize:           readSize,
+		rgsMin:             rgsMin,
+		rgsMax:             rgsMax,
+		filter:             filter,
+		curr:               EmptyRowNumber(),
+		at:                 at,
+		maxDefinitionLevel: maxDefinitionLevel,
 	}
 
 	// Apply options
@@ -1096,7 +1091,7 @@ func (c *SyncIterator) seekWithinPage(to RowNumber, definitionLevel int) {
 	pq.Release(c.currPage)
 	c.currPage = pg
 	c.currPageMin = c.curr
-	c.setValues(pg.Values())
+	c.currValues = pg.Values()
 	c.currPageN = 0
 	syncIteratorPoolPut(c.currBuf)
 	c.currBuf = nil
@@ -1149,16 +1144,7 @@ func (c *SyncIterator) next() (RowNumber, *pq.Value, error) {
 		}
 		if c.currBufN >= len(c.currBuf) || len(c.currBuf) == 0 {
 			c.currBuf = c.currBuf[:cap(c.currBuf)]
-
-			var n int
-			var err error
-
-			if c.currNestedValues != nil && c.filter != nil { // jpe - everything is terrible. clean up
-				n, err = c.currNestedValues.ReadNestedValues(c.currBuf)
-			} else {
-				n, err = c.currValues.ReadValues(c.currBuf)
-			}
-
+			n, err := c.currValues.ReadValues(c.currBuf)
 			if err != nil && !errors.Is(err, io.EOF) {
 				return EmptyRowNumber(), nil, err
 			}
@@ -1173,35 +1159,11 @@ func (c *SyncIterator) next() (RowNumber, *pq.Value, error) {
 
 		// Consume current buffer until empty
 		for c.currBufN < len(c.currBuf) {
-			var r, d byte
-			var v *pq.Value
-
-			// jpe - what about that one interaction where filter == nil, do we just do the old way?
-			if c.filter == nil {
-				v = &c.currBuf[c.currBufN]
-				r = byte(v.RepetitionLevel())
-				d = byte(v.DefinitionLevel())
-			} else if c.currNestedValues != nil { // jpe - better check
-
-				// if currNestedValues is set advance c.curr AS LONG AS the the definition level DOES NOT equal c.definitionLevel. in this case the value is a NULL and we can skip it
-				r = c.currPage.RepetitionLevels()[c.currPageN]
-				d = c.currPage.DefinitionLevels()[c.currPageN]
-				if d != c.defLvl {
-					// this is a null!
-					c.curr.Next(int(r), int(d))
-					c.currPageN++
-					continue
-				}
-				v = &c.currBuf[c.currBufN]
-			} else {
-				v = &c.currBuf[c.currBufN]
-				r = byte(v.RepetitionLevel())
-				d = byte(v.DefinitionLevel())
-			}
+			v := &c.currBuf[c.currBufN]
 
 			// Inspect all values to track the current row number,
 			// even if the value is filtered out next.
-			c.curr.Next(int(r), int(d))
+			c.curr.Next(v.RepetitionLevel(), v.DefinitionLevel(), c.maxDefinitionLevel)
 			c.currBufN++
 			c.currPageN++
 
@@ -1233,7 +1195,6 @@ func (c *SyncIterator) setPage(pg pq.Page) {
 
 	// Reset value buffers
 	c.currValues = nil
-	c.currNestedValues = nil
 	c.currPageMax = EmptyRowNumber()
 	c.currPageMin = EmptyRowNumber()
 	c.currBufN = 0
@@ -1253,18 +1214,7 @@ func (c *SyncIterator) setPage(pg pq.Page) {
 		c.currPage = pg
 		c.currPageMin = c.curr
 		c.currPageMax = rn
-
-		c.setValues(pg.Values())
-	}
-}
-
-func (c *SyncIterator) setValues(v pq.ValueReader) {
-	c.currValues = v
-
-	if _, ok := v.(NestedValuesReader); ok {
-		c.currNestedValues = v.(NestedValuesReader)
-	} else {
-		c.currNestedValues = nil
+		c.currValues = pg.Values()
 	}
 }
 
@@ -1313,11 +1263,12 @@ func (c *SyncIterator) Close() {
 // the optional predicate to each chunk, page, and value.  Results are read by calling
 // Next() until it returns nil.
 type ColumnIterator struct {
-	rgs      []pq.RowGroup
-	col      int
-	colName  string
-	filter   *InstrumentedPredicate
-	selectAs string
+	rgs                []pq.RowGroup
+	col                int
+	colName            string
+	filter             *InstrumentedPredicate
+	selectAs           string
+	maxDefinitionLevel int
 
 	// Row number to seek to, protected by mutex.
 	// Less allocs than storing in atomic.Value
@@ -1341,16 +1292,17 @@ type columnIteratorBuffer struct {
 	values     []pq.Value
 }
 
-func NewColumnIterator(ctx context.Context, rgs []pq.RowGroup, column int, columnName string, readSize int, filter Predicate, selectAs string) *ColumnIterator {
+func NewColumnIterator(ctx context.Context, rgs []pq.RowGroup, column int, columnName string, readSize int, filter Predicate, selectAs string, maxDefinitionLevel int) *ColumnIterator {
 	c := &ColumnIterator{
-		rgs:      rgs,
-		col:      column,
-		colName:  columnName,
-		filter:   &InstrumentedPredicate{Pred: filter},
-		selectAs: selectAs,
-		quit:     make(chan struct{}),
-		ch:       make(chan *columnIteratorBuffer, 1),
-		currN:    -1,
+		rgs:                rgs,
+		col:                column,
+		colName:            columnName,
+		filter:             &InstrumentedPredicate{Pred: filter},
+		selectAs:           selectAs,
+		quit:               make(chan struct{}),
+		ch:                 make(chan *columnIteratorBuffer, 1),
+		currN:              -1,
+		maxDefinitionLevel: maxDefinitionLevel,
 	}
 
 	c.iter = func() { c.iterate(ctx, readSize) }
@@ -1470,7 +1422,7 @@ func (c *ColumnIterator) iterate(ctx context.Context, readSize int) {
 
 								// We have to do this for all values (even if the
 								// value is excluded by the predicate)
-								rn.Next(v.RepetitionLevel(), v.DefinitionLevel())
+								rn.Next(v.RepetitionLevel(), v.DefinitionLevel(), c.maxDefinitionLevel)
 
 								if c.filter != nil {
 									if !c.filter.KeepValue(v) {