--- a/pkg/parquetquery/util.go
+++ b/pkg/parquetquery/util.go
@@ -6,23 +6,23 @@ import (
 	pq "github.com/parquet-go/parquet-go"
 )
 
-func GetColumnIndexByPath(pf *pq.File, s string) (index, depth int) {
+func GetColumnIndexByPath(pf *pq.File, s string) (index, depth, maxDef int) {
 	colSelector := strings.Split(s, ".")
 	n := pf.Root()
 	for len(colSelector) > 0 {
 		n = n.Column(colSelector[0])
 		if n == nil {
-			return -1, -1
+			return -1, -1, -1
 		}
 
 		colSelector = colSelector[1:]
 		depth++
 	}
 
-	return n.Index(), depth
+	return n.Index(), depth, n.MaxDefinitionLevel() // jpe - this is correct?
 }
 
 func HasColumn(pf *pq.File, s string) bool {
-	index, _ := GetColumnIndexByPath(pf, s)
+	index, _, _ := GetColumnIndexByPath(pf, s)
 	return index >= 0
 }