--- a/tempodb/encoding/vparquet2/block_search.go
+++ b/tempodb/encoding/vparquet2/block_search.go
@@ -363,7 +363,7 @@ func makeIterFunc(ctx context.Context, rgs []parquet.RowGroup, pf *parquet.File)
 		}
 
 		if async {
-			return pq.NewColumnIterator(ctx, rgs, index, name, 1000, predicate, selectAs)
+			return pq.NewColumnIterator(ctx, rgs, index, name, 1000, predicate, selectAs, maxDef)
 		}
 
 		var opts []pq.SyncIteratorOpt
@@ -371,7 +371,7 @@ func makeIterFunc(ctx context.Context, rgs []parquet.RowGroup, pf *parquet.File)
 			opts = append(opts, pq.SyncIteratorOptIntern())
 		}
 
-		return pq.NewSyncIterator(ctx, rgs, index, name, 1000, predicate, selectAs, byte(maxDef), opts...) // jpe - bytecast ok?
+		return pq.NewSyncIterator(ctx, rgs, index, name, 1000, predicate, selectAs, maxDef, opts...) // jpe - bytecast ok?
 	}
 }