… CHANGELOG.md
… cmd/tempo-cli/cmd-rewrite-blocks_test.go
… pkg/parquetquery/iters_test.go
… pkg/parquetquery/predicate_test.go
… tempodb/encoding/vparquet3/block_search.go
… tempodb/encoding/vparquet3/block_search_tags.go
… tempodb/encoding/vparquet3/wal_block.go
… tempodb/encoding/vparquet4/block_autocomplete.go
… tempodb/encoding/vparquet4/block_autocomplete_test.go
… tempodb/encoding/vparquet4/block_findtracebyid.go
… tempodb/encoding/vparquet4/block_findtracebyid_test.go
… tempodb/encoding/vparquet4/block_iterator.go
… tempodb/encoding/vparquet4/block_search.go
… tempodb/encoding/vparquet4/block_search_tags.go
… tempodb/encoding/vparquet4/block_search_tags_test.go
… tempodb/encoding/vparquet4/block_search_test.go
… tempodb/encoding/vparquet4/block_traceql_test.go
… tempodb/encoding/vparquet4/wal_block.go
… vendor/github.com/parquet-go/parquet-go/page_values.go