**What this PR does**:
Passes the max definition level into the `RowNumber.Next()` function to reduce the amount of required work. This allows for inlining and less work done when iterating at lower definition levels. This does mean that heavy iteration at higher definition levels is a bit slower. Putting up this PR b/c it shows a lot of promise, but I still want to test a version that reverts to the unrolled version for higher max definition levels.

This PR is a bit messy 😬. Look at `RowNumber.Next` (I can't link to it b/c it's in the collapsed iter file). All substantive changes flow from this. 

**Other Changes**:
- Refactored all benchmarks that work on local blocks to respect three env vars for easier testing. `BENCH_BLOCKID`, `BENCH_PATH`, `BENCH_TENANTID`
- Passed a limit of 20 into the traceql benchmark to better mimic real world behavior.
- Removed the unrolled `RowNumber.Next()` implementation although I may restore it.
- Added a `BenchmarkIterators` function to better test raw iterator combinations. I find it useful for testing, but am having issues translating perf improvements found here to TraceQL queries.

Generally I feel like more work could be done on our benchmarks to better guide us when making performance improvements. This is a first step. 

<details>
<summary>really long exhaustive benchmarks</summary>

```
goos: darwin
goarch: arm64
pkg: github.com/grafana/tempo/tempodb/encoding/vparquet4
cpu: Apple M3 Pro
                                                                                                           │ before.txt  │              after.txt              │
                                                                                                           │   sec/op    │   sec/op     vs base                │
FetchTagValues/tag:_span.http.url,_query:_{resource.namespace="tempo-ops"}-11                                14.30m ± 1%   14.15m ± 0%   -1.03% (p=0.000 n=10)
FetchTagValues/tag:_span.component,_query:_{resource.namespace="tempo-ops"}-11                               14.41m ± 2%   14.10m ± 1%   -2.12% (p=0.000 n=10)
FetchTagValues/tag:_span.http.url,_query:_{resource.namespace="tempo-ops"_&&_span.http.status_code=200}-11   14.40m ± 1%   14.17m ± 2%   -1.62% (p=0.009 n=10)
FetchTagValues/tag:_resource.namespace,_query:_{span.http.status_code=200}-11                                549.8m ± 0%   489.1m ± 1%  -11.04% (p=0.000 n=10)
FetchTags/query:_{resource.namespace="tempo-ops"}_span-11                                                    14.40m ± 1%   14.22m ± 1%   -1.23% (p=0.000 n=10)
FetchTags/query:_{resource.namespace="tempo-ops"}_resource-11                                                14.43m ± 0%   14.22m ± 1%   -1.49% (p=0.000 n=10)
FetchTags/query:_{resource.namespace="tempo-ops"}_none-11                                                    14.62m ± 1%   14.28m ± 1%   -2.33% (p=0.000 n=10)
FetchTags/query:_{resource.k8s.node.name>"h"}_span-11                                                         2.129 ± 0%    2.058 ± 1%   -3.30% (p=0.000 n=10)
FetchTags/query:_{resource.k8s.node.name>"h"}_resource-11                                                    947.8m ± 0%   908.9m ± 0%   -4.10% (p=0.000 n=10)
FetchTags/query:_{resource.k8s.node.name>"h"}_none-11                                                         8.772 ± 0%    8.533 ± 0%   -2.72% (p=0.000 n=10)
FetchTags/query:_{span.http.status_code=200}_span-11                                                          1.492 ± 0%    1.386 ± 0%   -7.15% (p=0.000 n=10)
FetchTags/query:_{span.http.status_code=200}_resource-11                                                     856.5m ± 0%   778.0m ± 0%   -9.17% (p=0.000 n=10)
FetchTags/query:_{span.http.status_code=200}_none-11                                                          4.624 ± 0%    4.330 ± 0%   -6.36% (p=0.000 n=10)
FetchTags/query:_{nestedSetParent=-1}_span-11                                                                 1.427 ± 0%    1.322 ± 0%   -7.41% (p=0.000 n=10)
FetchTags/query:_{nestedSetParent=-1}_resource-11                                                            947.9m ± 0%   868.1m ± 0%   -8.42% (p=0.000 n=10)
FetchTags/query:_{nestedSetParent=-1}_none-11                                                                 3.968 ± 0%    3.761 ± 0%   -5.21% (p=0.000 n=10)
FetchTags/query:_{rootName="Memcache.Put"}_span-11                                                           34.58m ± 0%   32.63m ± 0%   -5.64% (p=0.000 n=10)
FetchTags/query:_{rootName="Memcache.Put"}_resource-11                                                       14.90m ± 0%   14.41m ± 0%   -3.29% (p=0.000 n=10)
FetchTags/query:_{rootName="Memcache.Put"}_none-11                                                           149.3m ± 0%   145.8m ± 0%   -2.33% (p=0.000 n=10)
BackendBlockTraceQL/spanAttValMatch-11                                                                       84.91m ± 1%   79.34m ± 0%   -6.56% (p=0.000 n=10)
BackendBlockTraceQL/spanAttValNoMatch-11                                                                     5.589m ± 1%   5.389m ± 1%   -3.58% (p=0.000 n=10)
BackendBlockTraceQL/spanAttIntrinsicMatch-11                                                                 59.21m ± 0%   56.19m ± 0%   -5.11% (p=0.000 n=10)
BackendBlockTraceQL/spanAttIntrinsicNoMatch-11                                                               5.488m ± 1%   5.277m ± 0%   -3.86% (p=0.000 n=10)
BackendBlockTraceQL/resourceAttValMatch-11                                                                   380.8m ± 1%   368.4m ± 0%   -3.26% (p=0.000 n=10)
BackendBlockTraceQL/resourceAttValNoMatch-11                                                                 5.202m ± 0%   5.004m ± 0%   -3.81% (p=0.000 n=10)
BackendBlockTraceQL/resourceAttIntrinsicMatch-11                                                             24.96m ± 0%   24.38m ± 0%   -2.33% (p=0.000 n=10)
BackendBlockTraceQL/resourceAttIntrinsicMatch#01-11                                                          5.003m ± 1%   5.011m ± 0%        ~ (p=0.481 n=10)
BackendBlockTraceQL/traceOrMatch-11                                                                          226.8m ± 0%   222.7m ± 1%   -1.83% (p=0.000 n=10)
BackendBlockTraceQL/traceOrNoMatch-11                                                                        226.6m ± 0%   223.0m ± 0%   -1.59% (p=0.000 n=10)
BackendBlockTraceQL/mixedValNoMatch-11                                                                       173.2m ± 0%   170.2m ± 0%   -1.72% (p=0.000 n=10)
BackendBlockTraceQL/mixedValMixedMatchAnd-11                                                                 4.876m ± 3%   4.869m ± 0%        ~ (p=0.123 n=10)
BackendBlockTraceQL/mixedValMixedMatchOr-11                                                                  128.7m ± 0%   135.8m ± 0%   +5.54% (p=0.000 n=10)
BackendBlockTraceQL/count-11                                                                                 284.4m ± 0%   297.7m ± 1%   +4.68% (p=0.000 n=10)
BackendBlockTraceQL/struct-11                                                                                384.5m ± 2%   411.7m ± 4%   +7.07% (p=0.000 n=10)
BackendBlockTraceQL/||-11                                                                                    131.1m ± 0%   137.1m ± 0%   +4.61% (p=0.000 n=10)
BackendBlockTraceQL/mixed-11                                                                                 24.81m ± 1%   24.74m ± 0%   -0.29% (p=0.004 n=10)
BackendBlockTraceQL/complex-11                                                                               4.678m ± 1%   4.982m ± 1%   +6.50% (p=0.002 n=10)
BackendBlockTraceQL/select-11                                                                                4.713m ± 3%   4.989m ± 1%   +5.86% (p=0.002 n=10)
BackendBlockQueryRange/{}_|_rate()/5-11                                                                      818.5m ± 6%   850.7m ± 1%   +3.93% (p=0.043 n=10)
BackendBlockQueryRange/{}_|_rate()/7-11                                                                      854.2m ± 1%   859.7m ± 0%        ~ (p=0.353 n=10)
BackendBlockQueryRange/{}_|_rate()_by_(name)/5-11                                                             1.707 ± 3%    1.748 ± 5%        ~ (p=0.165 n=10)
BackendBlockQueryRange/{}_|_rate()_by_(name)/7-11                                                             1.879 ± 2%    1.931 ± 1%   +2.78% (p=0.001 n=10)
BackendBlockQueryRange/{}_|_rate()_by_(resource.service.name)/5-11                                            1.163 ± 8%    1.184 ± 7%        ~ (p=0.579 n=10)
BackendBlockQueryRange/{}_|_rate()_by_(resource.service.name)/7-11                                            1.267 ± 1%    1.299 ± 1%   +2.59% (p=0.000 n=10)
BackendBlockQueryRange/{}_|_rate()_by_(span.http.url)/5-11                                                    1.917 ± 2%    1.930 ± 1%        ~ (p=0.315 n=10)
BackendBlockQueryRange/{}_|_rate()_by_(span.http.url)/7-11                                                    1.996 ± 3%    2.028 ± 4%   +1.59% (p=0.009 n=10)
BackendBlockQueryRange/{resource.service.name=`loki-ingester`}_|_rate()/5-11                                 60.70m ± 1%   61.38m ± 0%   +1.12% (p=0.015 n=10)
BackendBlockQueryRange/{resource.service.name=`loki-ingester`}_|_rate()/7-11                                 61.07m ± 5%   61.54m ± 0%   +0.78% (p=0.023 n=10)
BackendBlockQueryRange/{status=error}_|_rate()/5-11                                                          86.87m ± 0%   80.88m ± 0%   -6.89% (p=0.000 n=10)
BackendBlockQueryRange/{status=error}_|_rate()/7-11                                                          87.38m ± 7%   80.97m ± 0%   -7.34% (p=0.000 n=10)
geomean                                                                                                      138.7m        136.3m        -1.72%

                                                                                                           │  before.txt   │               after.txt               │
                                                                                                           │     B/op      │      B/op       vs base               │
FetchTagValues/tag:_span.http.url,_query:_{resource.namespace="tempo-ops"}-11                                14.69Mi ±  1%    14.77Mi ±  2%       ~ (p=0.631 n=10)
FetchTagValues/tag:_span.component,_query:_{resource.namespace="tempo-ops"}-11                               14.63Mi ±  1%    14.71Mi ±  2%       ~ (p=0.971 n=10)
FetchTagValues/tag:_span.http.url,_query:_{resource.namespace="tempo-ops"_&&_span.http.status_code=200}-11   14.63Mi ±  2%    14.72Mi ±  2%       ~ (p=0.579 n=10)
FetchTagValues/tag:_resource.namespace,_query:_{span.http.status_code=200}-11                                8.268Mi ± 16%    7.467Mi ±  8%  -9.68% (p=0.035 n=10)
FetchTags/query:_{resource.namespace="tempo-ops"}_span-11                                                    14.69Mi ±  2%    14.67Mi ±  2%       ~ (p=0.971 n=10)
FetchTags/query:_{resource.namespace="tempo-ops"}_resource-11                                                14.72Mi ±  2%    14.75Mi ±  2%       ~ (p=0.796 n=10)
FetchTags/query:_{resource.namespace="tempo-ops"}_none-11                                                    14.74Mi ±  1%    14.68Mi ±  2%       ~ (p=0.481 n=10)
FetchTags/query:_{resource.k8s.node.name>"h"}_span-11                                                        25.91Mi ±  9%    25.57Mi ±  3%       ~ (p=0.436 n=10)
FetchTags/query:_{resource.k8s.node.name>"h"}_resource-11                                                    18.16Mi ±  2%    18.31Mi ±  3%       ~ (p=0.315 n=10)
FetchTags/query:_{resource.k8s.node.name>"h"}_none-11                                                        54.81Mi ± 22%    45.75Mi ± 22%       ~ (p=0.075 n=10)
FetchTags/query:_{span.http.status_code=200}_span-11                                                         12.78Mi ± 15%    13.40Mi ±  7%       ~ (p=0.165 n=10)
FetchTags/query:_{span.http.status_code=200}_resource-11                                                     6.697Mi ± 10%    7.097Mi ± 20%       ~ (p=0.190 n=10)
FetchTags/query:_{span.http.status_code=200}_none-11                                                         38.62Mi ± 34%    41.24Mi ± 23%       ~ (p=0.684 n=10)
FetchTags/query:_{nestedSetParent=-1}_span-11                                                                11.63Mi ± 18%    12.16Mi ± 27%       ~ (p=0.684 n=10)
FetchTags/query:_{nestedSetParent=-1}_resource-11                                                            9.454Mi ± 26%    9.880Mi ± 24%       ~ (p=0.739 n=10)
FetchTags/query:_{nestedSetParent=-1}_none-11                                                                42.20Mi ± 34%    40.34Mi ± 33%       ~ (p=0.529 n=10)
FetchTags/query:_{rootName="Memcache.Put"}_span-11                                                           7.232Mi ±  3%    7.121Mi ±  1%       ~ (p=0.052 n=10)
FetchTags/query:_{rootName="Memcache.Put"}_resource-11                                                       6.738Mi ±  1%    6.747Mi ±  1%       ~ (p=0.853 n=10)
FetchTags/query:_{rootName="Memcache.Put"}_none-11                                                           15.27Mi ± 12%    15.39Mi ± 22%       ~ (p=1.000 n=10)
BackendBlockTraceQL/spanAttValMatch-11                                                                       45.34Mi ±  1%    45.36Mi ±  1%       ~ (p=0.436 n=10)
BackendBlockTraceQL/spanAttValNoMatch-11                                                                     6.761Mi ±  2%    6.859Mi ±  1%  +1.44% (p=0.004 n=10)
BackendBlockTraceQL/spanAttIntrinsicMatch-11                                                                 39.76Mi ±  1%    39.66Mi ±  0%       ~ (p=0.063 n=10)
BackendBlockTraceQL/spanAttIntrinsicNoMatch-11                                                               7.078Mi ±  2%    7.054Mi ±  2%       ~ (p=0.529 n=10)
BackendBlockTraceQL/resourceAttValMatch-11                                                                   576.6Mi ±  0%    576.1Mi ±  0%       ~ (p=0.481 n=10)
BackendBlockTraceQL/resourceAttValNoMatch-11                                                                 5.514Mi ±  1%    5.349Mi ±  2%  -2.99% (p=0.000 n=10)
BackendBlockTraceQL/resourceAttIntrinsicMatch-11                                                             10.49Mi ±  2%    10.40Mi ±  1%       ~ (p=0.105 n=10)
BackendBlockTraceQL/resourceAttIntrinsicMatch#01-11                                                          6.833Mi ±  0%    6.864Mi ±  1%       ~ (p=0.143 n=10)
BackendBlockTraceQL/traceOrMatch-11                                                                          9.157Mi ± 18%    9.621Mi ± 24%       ~ (p=0.315 n=10)
BackendBlockTraceQL/traceOrNoMatch-11                                                                        9.226Mi ± 12%   10.402Mi ± 15%       ~ (p=0.123 n=10)
BackendBlockTraceQL/mixedValNoMatch-11                                                                       8.225Mi ±  9%    8.088Mi ± 13%       ~ (p=0.853 n=10)
BackendBlockTraceQL/mixedValMixedMatchAnd-11                                                                 5.976Mi ±  1%    5.945Mi ±  1%       ~ (p=0.218 n=10)
BackendBlockTraceQL/mixedValMixedMatchOr-11                                                                  7.859Mi ± 11%    8.189Mi ±  8%       ~ (p=0.684 n=10)
BackendBlockTraceQL/count-11                                                                                 418.3Mi ±  0%    417.8Mi ±  0%       ~ (p=0.105 n=10)
BackendBlockTraceQL/struct-11                                                                                14.30Mi ± 25%    13.95Mi ± 14%       ~ (p=1.000 n=10)
BackendBlockTraceQL/||-11                                                                                    16.65Mi ±  4%    16.55Mi ±  3%       ~ (p=1.000 n=10)
BackendBlockTraceQL/mixed-11                                                                                 6.942Mi ±  2%    6.949Mi ±  1%       ~ (p=0.631 n=10)
BackendBlockTraceQL/complex-11                                                                               6.150Mi ±  2%    5.803Mi ±  3%  -5.65% (p=0.000 n=10)
BackendBlockTraceQL/select-11                                                                                5.883Mi ±  3%    5.769Mi ±  1%  -1.94% (p=0.004 n=10)
BackendBlockQueryRange/{}_|_rate()/5-11                                                                      28.90Mi ±  9%    29.13Mi ±  7%       ~ (p=0.912 n=10)
BackendBlockQueryRange/{}_|_rate()/7-11                                                                      27.98Mi ± 24%    28.68Mi ± 23%       ~ (p=0.579 n=10)
BackendBlockQueryRange/{}_|_rate()_by_(name)/5-11                                                            32.83Mi ± 57%    32.89Mi ± 63%       ~ (p=0.739 n=10)
BackendBlockQueryRange/{}_|_rate()_by_(name)/7-11                                                            31.80Mi ± 43%    32.27Mi ± 11%       ~ (p=0.971 n=10)
BackendBlockQueryRange/{}_|_rate()_by_(resource.service.name)/5-11                                           31.94Mi ± 52%    29.54Mi ± 47%       ~ (p=0.353 n=10)
BackendBlockQueryRange/{}_|_rate()_by_(resource.service.name)/7-11                                           27.13Mi ± 43%    30.95Mi ± 12%       ~ (p=0.165 n=10)
BackendBlockQueryRange/{}_|_rate()_by_(span.http.url)/5-11                                                   355.0Mi ±  4%    360.7Mi ±  3%       ~ (p=0.165 n=10)
BackendBlockQueryRange/{}_|_rate()_by_(span.http.url)/7-11                                                   390.1Mi ±  3%    382.5Mi ±  3%  -1.95% (p=0.035 n=10)
BackendBlockQueryRange/{resource.service.name=`loki-ingester`}_|_rate()/5-11                                 23.61Mi ±  2%    23.07Mi ±  3%  -2.31% (p=0.043 n=10)
BackendBlockQueryRange/{resource.service.name=`loki-ingester`}_|_rate()/7-11                                 23.68Mi ±  4%    23.35Mi ±  2%       ~ (p=0.631 n=10)
BackendBlockQueryRange/{status=error}_|_rate()/5-11                                                          6.389Mi ±  3%    6.711Mi ±  5%       ~ (p=0.063 n=10)
BackendBlockQueryRange/{status=error}_|_rate()/7-11                                                          6.493Mi ±  3%    6.571Mi ±  6%       ~ (p=0.247 n=10)
geomean                                                                                                      18.17Mi          18.21Mi        +0.22%

                                                                                                           │  before.txt  │              after.txt              │
                                                                                                           │  allocs/op   │  allocs/op    vs base               │
FetchTagValues/tag:_span.http.url,_query:_{resource.namespace="tempo-ops"}-11                                81.37k ±  0%   81.38k ±  0%       ~ (p=0.469 n=10)
FetchTagValues/tag:_span.component,_query:_{resource.namespace="tempo-ops"}-11                               81.37k ±  0%   81.37k ±  0%       ~ (p=0.785 n=10)
FetchTagValues/tag:_span.http.url,_query:_{resource.namespace="tempo-ops"_&&_span.http.status_code=200}-11   81.39k ±  0%   81.39k ±  0%       ~ (p=0.324 n=10)
FetchTagValues/tag:_resource.namespace,_query:_{span.http.status_code=200}-11                                89.72k ±  0%   89.68k ±  0%  -0.05% (p=0.001 n=10)
FetchTags/query:_{resource.namespace="tempo-ops"}_span-11                                                    81.39k ±  0%   81.39k ±  0%       ~ (p=0.810 n=10)
FetchTags/query:_{resource.namespace="tempo-ops"}_resource-11                                                81.39k ±  0%   81.39k ±  0%       ~ (p=0.724 n=10)
…