@stoewer 2025-02-26
Looks good

> but I still want to test a version that reverts to the unrolled version for higher max definition levels

Definitely interested to see the benchmark results for this approach :+1: